refactor excel column helpers
